Resolving package split for geode-management
