+++ geode-assembly/src/distributedTest/java/org/apache/geode/management/client/ClusterManagementServiceBuilderDUnitTest.java
@@ -25,6 +25,7 @@ import org.springframework.web.client.RestTemplate;
 
 import org.apache.geode.management.api.ConnectionConfig;
 import org.apache.geode.management.api.RestTemplateClusterManagementServiceTransport;
+import org.apache.geode.management.cluster.client.ClusterManagementServiceBuilder;
 import org.apache.geode.test.dunit.rules.ClusterStartupRule;
 import org.apache.geode.test.dunit.rules.MemberVM;
 import org.apache.geode.test.junit.rules.MemberStarterRule;
