+++ geode-assembly/src/integrationTest/java/org/apache/geode/management/internal/rest/controllers/ClusterManagementRestLoggingTest.java
@@ -22,7 +22,7 @@ import org.junit.ClassRule;
 import org.junit.Test;
 
 import org.apache.geode.management.api.ClusterManagementService;
-import org.apache.geode.management.client.ClusterManagementServiceBuilder;
+import org.apache.geode.management.cluster.client.ClusterManagementServiceBuilder;
 import org.apache.geode.test.junit.rules.LocatorStarterRule;
 
 public class ClusterManagementRestLoggingTest {
