+++ geode-core/src/main/java/org/apache/geode/internal/config/VersionAdapter.java
