+++ geode-management/src/main/java/org/apache/geode/management/internal/dormant/Dormant.java
@@ -12,12 +12,11 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
+package org.apache.geode.management.internal.dormant;
 
-package org.apache.geode.management.internal;
-
-public interface Constants {
-  /**
-   * used by ClientCMS to request additional rest information to help with deserialization
-   */
-  String INCLUDE_CLASS_HEADER = "X-Include-Class";
+/**
+ * lazy-initialization interface
+ */
+public interface Dormant {
+  void wakeUp();
 }
