+++ geode-web-management/src/integrationTest/java/org/apache/geode/management/internal/rest/HateoasIntegrationTest.java
@@ -37,7 +37,7 @@ import org.apache.geode.management.api.ClusterManagementException;
 import org.apache.geode.management.api.ClusterManagementResult;
 import org.apache.geode.management.api.ClusterManagementService;
 import org.apache.geode.management.api.RestTemplateClusterManagementServiceTransport;
-import org.apache.geode.management.client.ClusterManagementServiceBuilder;
+import org.apache.geode.management.cluster.client.ClusterManagementServiceBuilder;
 import org.apache.geode.management.configuration.Region;
 import org.apache.geode.management.configuration.RegionType;
